refactor : Provider별 인증 로직 분리 및 전략 패턴 적용
변경 사항 :
- Kakao와 Apple 로그인 로직을 각각 KakaoAuthProviderHandler와 AppleAuthProviderHandler로 분리
- 공통 인터페이스(AuthProviderHandler)를 도입하여 Provider별 인증 처리 로직을 캡슐화함
- AuthService는 기존의 복잡한 로직 대신 각 핸들러를 호출하여 인증 처리를 위임하도록 리팩토링
- SRP 원칙을 준수하고, 확장성 및 유지보수성을 향상시킴 (비즈니스 로직 변경 없이 내부 구조만 개선)
