+++ clog-domain/src/main/kotlin/org/depromeet/clog/server/domain/auth/application/strategy/AuthProviderHandler.kt
@@ -1,0 +1,7 @@
+package org.depromeet.clog.server.domain.auth.application.strategy
+
+import org.depromeet.clog.server.domain.auth.application.dto.AuthResponseDto
+
+interface AuthProviderHandler<T> {
+    fun login(request: T): AuthResponseDto
+}
